+++ pytorch_lightning/trainer/callback_config.py
@@ -1,7 +1,5 @@
-import os
 from abc import ABC, abstractmethod
-from typing import List, Callable, Optional
-
+from typing import List, Optional
 
 from pytorch_lightning.callbacks import Callback, ModelCheckpoint, EarlyStopping, ProgressBarBase, ProgressBar
 from pytorch_lightning.loggers import LightningLoggerBase