+++ pytorch_lightning/trainer/distrib_data_parallel.py
@@ -142,9 +142,8 @@ from pkg_resources import parse_version
 import torch
 from pytorch_lightning import _logger as log
 from pytorch_lightning.loggers import LightningLoggerBase
-from pytorch_lightning.utilities import NATIVE_AMP_AVALAIBLE
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
-from pytorch_lightning.utilities.distributed import rank_zero_only, rank_zero_warn, rank_zero_info
+from pytorch_lightning.utilities.distributed import rank_zero_warn, rank_zero_info
 from pytorch_lightning.core.datamodule import LightningDataModule
 from pytorch_lightning.core.lightning import LightningModule
 
@@ -152,9 +151,7 @@ from pytorch_lightning.core.lightning import LightningModule
 try:
     from apex import amp
 except ImportError:
-    APEX_AVAILABLE = False
-else:
-    APEX_AVAILABLE = True
+    amp = None
 
 try:
     import horovod.torch as hvd
@@ -211,11 +208,6 @@ class TrainerDDPMixin(ABC):
     @property
     @abstractmethod
     def num_gpus(self) -> int:
-        """Warning: this is just empty shell for code implemented in other class."""
-
-    @property
-    @abstractmethod
-    def use_amp(self) -> bool:
         """Warning: this is just empty shell for code implemented in other class."""
 
     @abstractmethod